+++ Distiller-server/src/main/java/com/mapr/distiller/server/recordtypes/SlimThreadResourceRecord.java
@@ -140,7 +140,7 @@ public class SlimThreadResourceRecord extends Record {
 			this.pid = Integer.parseInt(line.split("\\s+", 2)[0]);
 			this.commandName = line.split("\\(", 2)[1].split("\\)", 2)[0];
 			this.cpuUsageTicks = new BigInteger(parts[11]).add(new BigInteger(parts[12]));
-			this.startTime = Integer.parseInt(parts[19]);
+			this.startTime = Long.parseLong(parts[19]);
 			this.iowaitTicks = new BigInteger(parts[39]);
 			
 			//Process io file
